final chnages to code

+++ src/main/java/com/example/jkpvt/UserManagement/Modules/ModulesDTO.java
@@ -16,7 +16,7 @@ public class ModulesDTO {
     private String moduleIcon;
     private List<ResourcesDTO> resources;
     private List<Integer> resourceIds; // This is a list resource ids
-    private Integer modelOrder;
+    private Long modelOrder;
     private Integer pageOffset;
     private Integer pageSize;
 }
